Move the airing time logic to the ext module
