+++ core/model/src/commonMain/kotlin/com/chesire/nekomp/core/model/AiringTime.kt
@@ -1,0 +1,10 @@
+package com.chesire.nekomp.core.model
+
+import kotlinx.datetime.DayOfWeek
+
+data class AiringTime(
+    val dayOfWeek: DayOfWeek,
+    val hour: Int,
+    val minute: Int,
+    val timeZone: String
+)
