+++ library/datasource/airing/src/commonMain/kotlin/com/chesire/nekomp/library/datasource/airing/AiringAnime.kt
@@ -1,9 +1,9 @@
 package com.chesire.nekomp.library.datasource.airing
 
+import com.chesire.nekomp.core.model.AiringTime
 import com.chesire.nekomp.core.model.Image
 import com.chesire.nekomp.core.model.Season
 import com.chesire.nekomp.core.model.Titles
-import kotlinx.datetime.DayOfWeek
 
 data class AiringAnime(
     val malId: Int,
@@ -14,11 +14,4 @@ data class AiringAnime(
     val season: Season,
     val year: Int,
     val airingTime: AiringTime?
-)
-
-data class AiringTime(
-    val dayOfWeek: DayOfWeek,
-    val hour: Int,
-    val minute: Int,
-    val timeZone: String
 )
